Fix data field widget so that it is read-only if there is a parent. (#26)
In sofa it is possible, to have both a parent in a data (so copying the value on change)
as well as the data to be not flag as read-only.

This is a very questionnable design that can be fixed either by:
- implementing in BaseData::isReadOnly so that it returns true of there is a parent
- handle the case in the UI, this avoid do change a core behavior in SOFA

I would prefer that we clarify and fix the expected behavior in BaseData. But until then
I would support we reflact that in the UI.